cms-admin: Move shared action log components to components folder
Promote UserCell, ActionChip, and DiffHeader into the shared
src/actionLog/components folder. The existing import paths crossed
the per-record / cross-entity boundary (ActionLogsGrid imported
UserCell from actionLogGrid/, the per-record location). Co-locating
them under components/ removes that asymmetry and prepares the way
for an upcoming EntityActionLog grid that will reuse the same cells.

Also translate the remaining German labels in ActionLogGrid and
ActionGridToolbar to English ("Datum" -> "Date", "Geändert von" ->
"Changed by", "Versionshistorie" -> "Version history", "Markierte
Versionen vergleichen" -> "Compare selected versions").
