+++ packages/admin/cms-admin/src/actionLog/components/actionChip/__stories__/ActionChip.stories.tsx
@@ -6,7 +6,7 @@ type Story = StoryObj<typeof ActionChip>;
 const meta: Meta<typeof ActionChip> = {
     component: ActionChip,
     tags: ["!autodocs"],
-    title: "Action log/Action logs/Action chip",
+    title: "Action log/Components/Action chip",
 };
 export default meta;
