+++ packages/admin/cms-admin/src/index.ts
@@ -5,7 +5,7 @@ export { ActionLogGrid, actionLogGridFragment } from "./actionLog/actionLogGrid/
 export { ActionLogsGrid, actionLogsGridFragment } from "./actionLog/actionLogs/actionLogsGrid/ActionLogsGrid";
 export { ActionLogsPage } from "./actionLog/actionLogs/actionLogsPage/ActionLogsPage";
 export { ActionLogShowVersion, actionLogShowVersionFragment } from "./actionLog/actionLogShowVersion/ActionLogShowVersion";
-export { DiffHeader, type DiffHeaderProps } from "./actionLog/components/DiffHeader";
+export { DiffHeader, type DiffHeaderProps } from "./actionLog/components/diffHeader/DiffHeader";
 export { DiffViewer, type DiffViewerProps } from "./actionLog/components/diffViewer/DiffViewer";
 export { ActionLogHeader, type ActionLogHeaderProps } from "./actionLog/components/header/ActionLogHeader";
 export { AnchorBlock } from "./blocks/AnchorBlock";
